+++ tests/testthat/test-filter.r
@@ -231,3 +231,8 @@ test_that("filter correctly handles empty data frames (#782)", {
   expect_true( is.null(names(res)) )
 })
 
+test_that("filter(.,TRUE,TRUE) works (#1210)", {
+  df <- data.frame(x=1:5)
+  res <- filter(df,TRUE,TRUE)
+  
+})