+++ NEWS.md
@@ -12,6 +12,8 @@
 
 * hybrid evaluation does not take place for objects with a class (#1237). 
 
+* `filter` handles series of `TRUE` (#1210). 
+
 # dplyr 0.4.2
 
 This is a minor release containing fixes for a number of crashes and issues identified by R CMD CHECK. There is one new "feature": dplyr no longer complains about unrecognised attributes, and instead just copies them over to the output.